update userid db entry type

+++ bot/db/models.py
@@ -13,10 +13,11 @@ class User(Base):
     blacklisted = Column(Integer, default=0)
     
 class PrivateMessages(Base):
+    
     __tablename__ = 'messages'
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BIGINT, primary_key=True, autoincrement=True)
     username = Column(String(50), nullable=False)
-    userId = Column(Integer, ForeignKey('users.id'), nullable=False)
+    userId = Column(BIGINT, ForeignKey('users.id'), nullable=False)
     message = Column(Text, nullable=False)  # Changed to Text type
     createdAt = Column(DateTime, default=datetime.now)  # Use DateTime type for actual date-time objects
 
@@ -29,7 +30,7 @@ class ChatMessages(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     chatId = Column(BIGINT, nullable=False)
     username = Column(String(50), nullable=False)
-    userId = Column(Integer, ForeignKey('users.id'), nullable=False)
+    userId = Column(BIGINT, ForeignKey('users.id'), nullable=False)
     message = Column(Text, nullable=False)  # Changed to Text type
     createdAt = Column(DateTime, default=datetime.now)  # Use DateTime type for actual date-time objects
